chore: 공통 컴포넌트 packages/ui로 디렉토리 변경
